hw2_ch: start explicit size decode

+++ multiple_instructions_and_suffix/decoder/util.go
@@ -5,12 +5,12 @@ import (
 	"encoding/binary"
 )
 
-func Convert_16bits_to_int(a []byte) int16 {
+func Convert_16bits_to_int(a []byte) int {
 	var ret int16
 	buffer := bytes.NewReader(a)
 	err := binary.Read(buffer, binary.NativeEndian, &ret)
 	if err != nil {
 		panic("Failed read 2bytes into a 16bit binary")
 	}
-	return ret
+	return int(ret)
 }
